[AN] style: 코드 포맷팅 #111
