+++ android/app/src/main/java/com/onair/hearit/domain/Mapper.kt
@@ -1,14 +1,5 @@
 package com.onair.hearit.domain
 
-data class RandomHearitItem(
-    val id: Long,
-    val title: String,
-    val summary: String,
-    val source: String,
-    val playTime: Int,
-    val createdAt: String,
-)
-
 fun RandomHearitItem.toHearitShortsItem(
     audioUrl: String,
     script: List<ScriptLine>,
