+++ android/app/src/main/java/com/onair/hearit/domain/RandomHearitItem.kt
@@ -1,0 +1,10 @@
+package com.onair.hearit.domain
+
+data class RandomHearitItem(
+    val id: Long,
+    val title: String,
+    val summary: String,
+    val source: String,
+    val playTime: Int,
+    val createdAt: String,
+)
